Abstract base appearence and hover update from drag handles
